SYNC-290:  Sync Statistics page throws an error upon loading

+++ api/src/main/java/org/openmrs/module/sync/api/SyncService.java
@@ -306,7 +306,7 @@ public interface SyncService {
 	 * @throws APIException
 	 */
 	@Transactional(readOnly = true)
-	public Integer getCountOfSyncRecords(RemoteServer server, Date from, Date to, SyncRecordState... states)
+	public Long getCountOfSyncRecords(RemoteServer server, Date from, Date to, SyncRecordState... states)
 	                                                                                                        throws APIException;
 	
 	/**
