+++ api/src/main/java/org/openmrs/module/sync/api/impl/SyncServiceImpl.java
@@ -883,7 +883,7 @@ public class SyncServiceImpl implements SyncService {
 		return;
 	}
 	
-	public Integer getCountOfSyncRecords(RemoteServer server, Date from, Date to, SyncRecordState... states)
+	public Long getCountOfSyncRecords(RemoteServer server, Date from, Date to, SyncRecordState... states)
 	                                                                                                        throws APIException {
 		return dao.getCountOfSyncRecords(server, from, to, states);
 	}
